Move all plugin related modules to `Protocols.Plugin`
The separation between plugin related code and protocol related code
should be obvious from the module hierarchy
