+++ clash-protocols-base/clash-protocols-base.cabal
@@ -111,15 +111,16 @@ library
     , template-haskell
 
   exposed-modules:
-    Protocols.Circuit
-    Protocols.Cpp
-    Protocols.Internal.TaggedBundle
-    Protocols.Internal.TaggedBundle.TH
-    Protocols.Internal.Types
-    Protocols.Internal.Units
-    Protocols.Internal.Units.TH
     Protocols.Plugin
+    Protocols.Plugin.Cpp
     Protocols.Plugin.Internal
-    Protocols.Protocol.TH
+    Protocols.Plugin.TaggedBundle
+    Protocols.Plugin.TaggedBundle.TH
+    Protocols.Plugin.TH
+    Protocols.Plugin.Units
+    Protocols.Plugin.Units.TH
+
+  other-modules:
+    Protocols.Plugin.Types
 
   default-language: Haskell2010
