+++ clash-protocols-base/src/Protocols/Plugin.hs
@@ -5,6 +5,9 @@ A GHC source plugin providing a DSL for writing Circuit components. Credits to
 @circuit-notation@ at <https://github.com/cchalmers/circuit-notation>.
 -}
 module Protocols.Plugin (
+  Circuit (..),
+  Protocol (..),
+  CSignal,
   plugin,
   circuit,
   (-<),
@@ -13,11 +16,16 @@ module Protocols.Plugin (
 -- base
 import Prelude
 
+-- clash-prelude
+import qualified Clash.Explicit.Prelude as C
+
 -- clash-protocols
-import Protocols.Internal.Types
-import Protocols.Internal.TaggedBundle
-import Protocols.Internal.Units
+import Protocols.Plugin.Cpp
 import Protocols.Plugin.Internal
+import Protocols.Plugin.TH
+import Protocols.Plugin.TaggedBundle
+import Protocols.Plugin.Types
+import Protocols.Plugin.Units
 
 -- circuit-notation
 import qualified CircuitNotation as CN
@@ -27,6 +35,31 @@ import Data.Tagged
 
 -- ghc
 import qualified GHC.Plugins as GHC
+
+instance Protocol () where
+  type Fwd () = ()
+  type Bwd () = ()
+
+{- | __NB__: The documentation only shows instances up to /3/-tuples. By
+default, instances up to and including /12/-tuples will exist. If the flag
+@large-tuples@ is set instances up to the GHC imposed limit will exist. The
+GHC imposed limit is either 62 or 64 depending on the GHC version.
+-}
+instance Protocol (a, b) where
+  type Fwd (a, b) = (Fwd a, Fwd b)
+  type Bwd (a, b) = (Bwd a, Bwd b)
+
+-- Generate n-tuple instances, where n > 2
+protocolTupleInstances 3 maxTupleSize
+
+instance (C.KnownNat n) => Protocol (C.Vec n a) where
+  type Fwd (C.Vec n a) = C.Vec n (Fwd a)
+  type Bwd (C.Vec n a) = C.Vec n (Bwd a)
+
+-- XXX: Type families with Signals on LHS are currently broken on Clash:
+instance Protocol (CSignal dom a) where
+  type Fwd (CSignal dom a) = C.Signal dom a
+  type Bwd (CSignal dom a) = C.Signal dom ()
 
 -- | @circuit-notation@ plugin repurposed for "Protocols".
 plugin :: GHC.Plugin
