+++ clash-protocols-base/src/Protocols/Plugin/Units/TH.hs
@@ -1,6 +1,6 @@
 {-# OPTIONS_HADDOCK hide #-}
 
-module Protocols.Internal.Units.TH (unitsTupleInstances) where
+module Protocols.Plugin.Units.TH (unitsTupleInstances) where
 
 import Language.Haskell.TH
