+++ clash-protocols/src/Protocols/Axi4/Stream.hs
@@ -27,6 +27,7 @@ import qualified Clash.Prelude as C
 import qualified Protocols.Df as Df
 import qualified Protocols.DfConv as DfConv
 import Protocols.Hedgehog.Internal
+import Protocols.Idle
 import Protocols.Internal
 
 instance (KnownNat n) => Hashable (Unsigned n)
