+++ clash-protocols/src/Protocols/Axi4/WriteAddress.hs
@@ -47,6 +47,7 @@ import qualified Clash.Prelude as C
 
 -- me
 import Protocols.Axi4.Common
+import Protocols.Idle
 import Protocols.Internal
 
 -- | Configuration options for 'Axi4WriteAddress'.
