+++ clash-protocols/src/Protocols/Df.hs
@@ -129,6 +129,7 @@ import qualified Clash.Prelude as C
 import Clash.Signal.Internal (Signal (..))
 
 -- me
+import Protocols.Idle
 import Protocols.Internal
 
 {-# ANN module "HLint: ignore Use const" #-}
