+++ clash-protocols/src/Protocols/Idle.hs
@@ -4,24 +4,30 @@
 Functionalities to easily create idle circuits for protocols.
 -}
 module Protocols.Idle (
+  -- * Type classes
   IdleCircuit (..),
+
+  -- * Utility functions
   idleSource,
   idleSink,
+  forceResetSanityGeneric,
 ) where
 
-import qualified Clash.Prelude as C
+import Clash.Prelude
+import Prelude ()
+
 import Data.Proxy
-import Protocols.Cpp (maxTupleSize)
 import Protocols.Internal
 import Protocols.Internal.TH (idleCircuitTupleInstances)
+import Protocols.Plugin.Cpp (maxTupleSize)
 
 instance (IdleCircuit a, IdleCircuit b) => IdleCircuit (a, b) where
   idleFwd _ = (idleFwd $ Proxy @a, idleFwd $ Proxy @b)
   idleBwd _ = (idleBwd $ Proxy @a, idleBwd $ Proxy @b)
 
-instance (IdleCircuit a, C.KnownNat n) => IdleCircuit (C.Vec n a) where
-  idleFwd _ = C.repeat $ idleFwd $ Proxy @a
-  idleBwd _ = C.repeat $ idleBwd $ Proxy @a
+instance (IdleCircuit a, KnownNat n) => IdleCircuit (Vec n a) where
+  idleFwd _ = repeat $ idleFwd $ Proxy @a
+  idleBwd _ = repeat $ idleBwd $ Proxy @a
 
 instance IdleCircuit () where
   idleFwd _ = ()
@@ -37,3 +43,30 @@ idleSource = Circuit $ const ((), idleFwd $ Proxy @p)
 -- | Idle state of a sink, this circuit does not consume any data.
 idleSink :: forall p. (IdleCircuit p) => Circuit p ()
 idleSink = Circuit $ const (idleBwd $ Proxy @p, ())
+
+{- | Force a /nack/ on the backward channel and /no data/ on the forward
+channel if reset is asserted.
+-}
+forceResetSanityGeneric ::
+  forall dom a fwd bwd.
+  ( KnownDomain dom
+  , HiddenReset dom
+  , IdleCircuit a
+  , Fwd a ~ Signal dom fwd
+  , Bwd a ~ Signal dom bwd
+  ) =>
+  Circuit a a
+forceResetSanityGeneric = Circuit go
+ where
+  go (fwd, bwd) =
+    unbundle
+      $ mux
+        rstAsserted
+        (bundle (idleBwd $ Proxy @a, idleFwd $ Proxy @a))
+        (bundle (bwd, fwd))
+
+#if MIN_VERSION_clash_prelude(1,8,0)
+  rstAsserted = unsafeToActiveHigh hasReset
+#else
+  rstAsserted = unsafeToHighPolarity hasReset
+#endif
