+++ clash-protocols/src/Protocols/Internal/TH.hs
@@ -4,7 +4,6 @@ module Protocols.Internal.TH where
 
 import Control.Monad.Extra (concatMapM)
 import Language.Haskell.TH
-
 import Protocols.Internal.Types
 
 {- | Template haskell function to generate IdleCircuit instances for the tuples
